backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
python-socketio==5.8.0
gunicorn==20.1.0
gevent>=1.4.0
requests
python-dotenv
dnspython==1.15.0
flask-jwt-extended==4.4.4
Flask-Migrate==4.0.4